refactor: split dt_fitting into a per-stage package
dt_fitting.py had grown to 1691 lines against the 1000-line module cap.
The split is mechanical and behavior-preserving: constants, coarse
(rasterization / integer search / polarity), weights (Tukey and the
information-matrix inversion), transforms (pose math and normal
equations), ridge, lm, and basin, with the package __init__ re-exporting
the whole surface so consumers keep importing from
spindoctor.nav_technique.dt_fitting.  Largest module is now 611 lines.

The LM driver calls the ridge stage through the module object so the
stage stays monkeypatchable; the one test that patches it is repointed.
